+++ inference_core/src/lib.rs
@@ -3,3 +3,7 @@ pub mod hf;
 pub mod modelpool;
 pub mod models;
 pub mod tokenizer;
+pub(crate) mod generate;
+
+
+pub use generate::generate;
